DB Connector
Here, I created three strings to hold my sensitive data, then I developed a public class to load the properties from my config.properties file using FileInputStream, as it contains the login information for my database. This approach avoids the bad practice of leaving login information in the code. The process is enclosed within a try-catch block, which ensures the stream is closed after execution. My config.properties file is not on my GitHub because I used .gitignore, as the information is private, so I will upload the file with the assignment submission.