Lecturer Report Generator
For this lecturer report, the SQL query for the lectures table is aliased as 'l' to make the query clearer. Additionally, I used CONCAT and DISTINCT to group each course the lecturer can teach, ensuring that each course type appears only once due to the DISTINCT keyword. I also used a COUNT to determine the number of students in each module. I utilized a LEFT JOIN to associate each lecturer with the course they teach and determine the number of students enrolled in each course and everything is sorted by the LecturerID.
After running the query and retrieving the data, the result will be appended to the lecturer report. This report will contain the lecturer's name, role, teaching module, number of students in the module, and the type of courses they can teach.