Generate & Export
Here, I have created a method that uses a scanner to retrieve the user's option and determine the type of report to be generated using a switch statement. Once the option is selected, another question will be asked to the user, inquiring whether they wish to export the report. The scanner will retrieve a string input, either 'Y' or 'N', and store the value in the variable 'exportOption'. This value will then be used in an 'IF/ELSE' statement: if the user selects 'Y', my 'EXPORTREPORT' method will be executed.