Export Report Options
I designed this 'exportReport' method to obtain the user's choice using a scanner, which stores an integer value in the variable 'formatOption'. This value is then utilized in another switch statement to determine the format in which the user wants the report to be generated: either as CSV, TXT, or printed on the console. The writing methods will be implemented in 'exportToTXT' and 'exportToCSV'. Additionally, I included a fixed set of options using an enum called 'ReportType' to make the code clear and readable and to be used in the main method.