User Class
For the User class, I chose to use an abstract class so that other classes can utilize its information instead of directly creating a user object in other classes. Then, I protected the string information along with the UserRole so that it can only be accessed inside the User class,  protecting the username and password from being accessed from outside of my user class. I created a constructor so that whenever a class uses the User information, it will initialize with a username, password, and role.
Additionally, I added getter and setter methods to retrieve user information and also to set and change new users in the system.
The generateReport method in this class ensures that depending on which role is using the system, it will have its own specifications for generating the report. Because it is abstract, it will adapt to different roles.
The 'changeCredentials' method will use setter methods to update new usernames and passwords in the system.