Office Class
For the Office Class, I used a public class to ensure it can be accessed from other parts of the code. Additionally, I used 'extends' to make Office a subclass of my User class, ensuring that it will inherit the methods from the User class.
In the constructor of this class, when used, it is necessary to provide a username and password. The 'super'  will then call the constructor of the User class, passing to it the username, password, and because it is in the Office class, the role using the system.
The '@Override' annotation will modify my 'generateReport' method in the User class. However, because the logic to generate the reports is in my ReportsGenerator class, it will only display a message to let the user know that the reports are being generated.