UserAuthenticator Class
For my 'UserAuthenticator' class, I am using a 'Map' as a '<key-value>' pair structure to find the user associated with a given username. Then, a scanner is used to obtain the user's input. If the information matches the user data in the map, it will return the user,  otherwise, it won't authenticate the user.