Admin Class
I started by applying the same logic implemented in the Office and Lecturer classes and also, to prevent the admin from generating reports, I implemented the necessary logic in my main code. Additionally, I implemented a method that allows the admin to add users to the map <String, User> only if the user is not already in the system. The switch statement then executes the appropriate code based on the role.