UserModify Method
The modifyUser method first checks if the username entered by the admin in the menu, exists in the system. If it doesn't, the modification cannot proceed, and a message indicating that the user doesn't exist will be displayed. Additionally, another check is performed to verify if the new username is already in the system. In that case, it will request the selection of a unique username.
If both checks pass, the function .remove() will eliminate the 'oldUsername' from the map to avoid issues when updating the map, especially if a username is being updated. Then, using a setter, it will update the system with the new username and password, and if the role is changing, the method will instantiate a new user object of the appropriate class based on 'newRole'. This ensures that a new role is assigned to the new user. After all modifications are completed, .put() will re-add the user and their information back to the map.