DeleteUser Method
For the deleteUser, I created a check to ensure the username exists in the system. If there is such a username, the function .remove() will take the username and remove it from the map. Once the operation is completed, it will display a message saying that the user has been deleted successfully.