Switching between menus
Here, as the admin menu has to start first in the system, I implemented a logout method that will ask the admin if they want to logout after their actions on the system. If the answer is 'No', the admin menu will remain active to perform admin tasks. If they choose to logout, the login screen will appear, allowing other users added by the admin to log into the system, and the reports menu will open.
Additionally, I implemented a try-and-catch block in cases where the admin has to add or update the role of the staff. This is to handle situations where a role that is not available in the code doesn't crash the system.